multimodal_rag/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
jinja2>=3.1.2

# Vector Database & Embeddings
chromadb>=0.4.18
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.1.0
torchvision>=0.16.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
pillow>=10.1.0
pytesseract>=0.3.10
opencv-python>=4.8.0
easyocr>=1.7.0
python-pptx>=0.6.21

# Audio Processing
openai-whisper>=20231117
pydub>=0.25.1

# ML & Data Processing
numpy>=1.24.0
scikit-learn>=1.3.2
pandas>=2.1.0

# LLM Integration
httpx>=0.25.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
chardet>=5.2.0

# Language Detection & Translation
langdetect>=1.0.9
googletrans==4.0.0rc1

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.12.0
